refactor: organized into inference engine folders. Still WIP, but this runs (is close to the ideal end state) and does not break existing code.
