+++ type_infer/__init__.py
@@ -1,10 +1,10 @@
 from type_infer import base
 from type_infer import dtype
-from type_infer import infer
+from type_infer import api
 from type_infer import helpers
 
 
-__version__ = '0.0.17'
+__version__ = '0.0.18'
 
 
-__all__ = ['base', 'dtype', 'infer', 'helpers', '__version__']
+__all__ = ['base', 'dtype', 'api.py', 'helpers', '__version__']
