+++ type_infer/bert/infer.py
@@ -1,0 +1,1 @@
+STABLE = False
