+++ type_infer/dtype.py
@@ -45,3 +45,5 @@ class dtype:
     # Misc (Unk/NaNs)
     empty = "empty"
     invalid = "invalid"
+
+# TODO: introduce "modifiers"?
